[REFACT] 패키지 구조 적용
