--- src/main/java/com/prgrms/be/domain/user/application/RedisService.java
@@ -1,34 +1,0 @@
-package com.prgrms.be.domain.user.application;
-
-import java.time.Duration;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.ValueOperations;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-
-@Component
-@RequiredArgsConstructor
-@Transactional
-public class RedisService {
-
-    private final RedisTemplate<String, Object> redisTemplate;
-
-    public void setValues(String key, String data, Duration duration) {
-        ValueOperations<String, Object> values = redisTemplate.opsForValue();
-        values.set(key, data, duration);
-    }
-
-    @Transactional(readOnly = true)
-    public String getValues(String key) {
-        ValueOperations<String, Object> values = redisTemplate.opsForValue();
-        if (values.get(key) == null) {
-            return "false";
-        }
-        return (String) values.get(key);
-    }
-
-    protected boolean checkExistsValue(String value) {
-        return !value.equals("false");
-    }
-}
