--- src/main/java/com/prgrms/be/domain/user/domain/enums/SocialType.java
@@ -1,5 +1,0 @@
-package com.prgrms.be.domain.user.domain.enums;
-
-public enum SocialType {
-    GOOGLE, KAKAO, NAVER
-}
