--- src/main/java/com/prgrms/be/domain/user/infrastructure/UserJPARepository.java
@@ -1,17 +1,0 @@
-package com.prgrms.be.domain.user.infrastructure;
-
-import com.prgrms.be.domain.user.domain.entity.User;
-import com.prgrms.be.domain.user.domain.enums.SocialType;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserJPARepository extends JpaRepository<User, Long> {
-
-    Optional<User> findBySocialTypeAndSocialId(SocialType socialType, String socialId);
-
-    Optional<User> findByEmail(String email);
-
-    Optional<User> findByNickname(String nickname);
-
-    Optional<User> findByRefreshToken(String refreshToken);
-}
