--- src/main/java/com/prgrms/be/domain/user/jwt/service/JwtService.java
@@ -1,123 +1,0 @@
-package com.prgrms.be.domain.user.jwt.service;
-
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.algorithms.Algorithm;
-import com.prgrms.be.domain.user.application.RedisService;
-import com.prgrms.be.domain.user.infrastructure.UserJPARepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.time.Duration;
-import java.util.Date;
-import java.util.Optional;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-@Getter
-public class JwtService {
-
-    @Value("${jwt.secretKey}")
-    private String secretKey;
-
-    @Value("${jwt.access.expiration}")
-    private Long accessTokenExpirationPeriod;
-
-    @Value("${jwt.refresh.expiration}")
-    private Long refreshTokenExpirationPeriod;
-
-    @Value("${jwt.access.header}")
-    private String accessHeader;
-
-    @Value("${jwt.refresh.header}")
-    private String refreshHeader;
-
-    private static final String ACCESS_TOKEN_SUBJECT = "AccessToken";
-    private static final String REFRESH_TOKEN_SUBJECT = "RefreshToken";
-    private static final String EMAIL_CLAIM = "email";
-    private static final String BEARER = "Bearer ";
-
-    private final UserJPARepository userJPARepository;
-    private final RedisService redisService;
-
-    public String createAccessToken(String email) {
-        Date now = new Date();
-
-        return JWT.create()
-            .withSubject(ACCESS_TOKEN_SUBJECT)
-            .withExpiresAt(new Date(now.getTime() + accessTokenExpirationPeriod))
-            .withClaim(EMAIL_CLAIM, email)
-            .sign(Algorithm.HMAC512(secretKey));
-    }
-
-    public String createRefreshToken() {
-        Date now = new Date();
-        return JWT.create()
-            .withSubject(REFRESH_TOKEN_SUBJECT)
-            .withExpiresAt(new Date(now.getTime() + refreshTokenExpirationPeriod))
-            .sign(Algorithm.HMAC512(secretKey));
-    }
-
-    public void sendAccessToken(HttpServletResponse response, String accessToken) {
-        response.setStatus(HttpServletResponse.SC_OK);
-
-        setAccessTokenHeader(response, BEARER + accessToken);
-    }
-
-    public void sendAccessTokenAndRefreshToken(HttpServletResponse response, String accessToken,
-        String refreshToken) {
-        response.setStatus(HttpServletResponse.SC_OK);
-
-        setAccessTokenHeader(response, BEARER + accessToken);
-        setRefreshTokenHeader(response, BEARER + refreshToken);
-    }
-
-    public Optional<String> extractRefreshToken(HttpServletRequest request) {
-        return Optional.ofNullable(request.getHeader(refreshHeader))
-            .filter(refreshToken -> refreshToken.startsWith(BEARER))
-            .map(refreshToken -> refreshToken.replace(BEARER, ""));
-    }
-
-    public Optional<String> extractAccessToken(HttpServletRequest request) {
-        return Optional.ofNullable(request.getHeader(accessHeader))
-            .filter(accessToken -> accessToken.startsWith(BEARER))
-            .map(accessToken -> accessToken.replace(BEARER, ""));
-    }
-
-    public Optional<String> extractEmail(String accessToken) {
-        try {
-            return Optional.ofNullable(JWT.require(Algorithm.HMAC512(secretKey))
-                .build()
-                .verify(accessToken) //검증
-                .getClaim(EMAIL_CLAIM) //추출
-                .asString());
-        } catch (Exception e) { //JWTVerificationException, ClaimNotFoundException
-            return Optional.empty();
-        }
-    }
-
-    //RefreshToken redis 저장
-    public void updateRefreshToken(String email, String refreshToken) {
-        redisService.setValues(refreshToken, email,
-            Duration.ofMillis(refreshTokenExpirationPeriod));
-    }
-
-    public boolean isTokenValid(String token) {
-        try {
-            JWT.require(Algorithm.HMAC512(secretKey)).build().verify(token);
-            return true;
-        } catch (Exception e) { //JWTVerificationException
-            return false;
-        }
-    }
-
-    private void setAccessTokenHeader(HttpServletResponse response, String accessToken) {
-        response.setHeader(accessHeader, accessToken);
-    }
-
-    private void setRefreshTokenHeader(HttpServletResponse response, String refreshToken) {
-        response.setHeader(refreshHeader, refreshToken);
-    }
-}
