+++ src/main/java/com/prgrms/zzalmyu/domain/user/domain/entity/User.java
@@ -1,4 +1,62 @@
 package com.prgrms.zzalmyu.domain.user.domain.entity;
 
+import com.prgrms.zzalmyu.domain.user.domain.enums.Role;
+import com.prgrms.zzalmyu.domain.user.domain.enums.SocialType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import jakarta.validation.constraints.Email;
+import java.time.LocalDateTime;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Entity
+@Table(name = "user_TB")
 public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    @Email
+    private String email;
+
+    @Column(nullable = false, length = 10)
+    private String nickname;
+
+    @Enumerated(EnumType.STRING)
+    private Role role;
+
+    @Enumerated(EnumType.STRING)
+    private SocialType socialType;
+
+    private String socialId;
+
+    @CreationTimestamp
+    @Column(nullable = false)
+    private LocalDateTime createdAt;
+
+    private LocalDateTime deletedAt;
+
+    @Builder
+    private User(String email, String nickname, Role role, String socialId, SocialType socialType) {
+        this.email = email;
+        this.nickname = nickname;
+        this.role = role;
+        this.socialId = socialId;
+        this.socialType = socialType;
+    }
 }
