+++ src/main/java/com/prgrms/zzalmyu/domain/user/domain/enums/SocialType.java
@@ -1,0 +1,5 @@
+package com.prgrms.zzalmyu.domain.user.domain.enums;
+
+public enum SocialType {
+    GOOGLE, KAKAO, NAVER
+}
