+++ src/main/java/com/prgrms/zzalmyu/domain/user/infrastructure/UserJPARepository.java
@@ -1,0 +1,13 @@
+package com.prgrms.zzalmyu.domain.user.infrastructure;
+
+import com.prgrms.zzalmyu.domain.user.domain.entity.User;
+import com.prgrms.zzalmyu.domain.user.domain.enums.SocialType;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserJPARepository extends JpaRepository<User, Long> {
+
+    Optional<User> findBySocialTypeAndSocialId(SocialType socialType, String socialId);
+
+    Optional<User> findByEmail(String email);
+}
