+++ src/main/java/com/prgrms/zzalmyu/domain/user/oauth2/userinfo/GoogleOAuth2UserInfo.java
@@ -1,0 +1,40 @@
+package com.prgrms.zzalmyu.domain.user.oauth2.userinfo;
+
+import java.util.Map;
+
+/*
+    Google 유저정보 response 예시
+    {
+   "sub": "식별값",
+   "name": "name",
+   "given_name": "given_name",
+   "picture": "https//lh3.googleusercontent.com/~~",
+   "email": "email",
+   "email_verified": true,
+   "locale": "ko"
+}
+}
+}
+ */
+public class GoogleOAuth2UserInfo extends
+    OAuth2UserInfo {
+
+    public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
+        super(attributes);
+    }
+
+    @Override
+    public String getId() {
+        return (String) attributes.get("sub");
+    }
+
+    @Override
+    public String getNickname() {
+        return (String) attributes.get("name");
+    }
+
+    @Override
+    public String getEmail() {
+        return (String) attributes.get("email");
+    }
+}
